# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "De'Corrinna API",
        "description": "REST API backend De'Corrinna (Laravel 12 + PostgreSQL). Autentikasi memakai Bearer token (Laravel Sanctum).",
        "contact": {
            "name": "Tim Azatech",
            "email": "dev@azatech.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server lokal (dev)"
        }
    ],
    "paths": {
...
        "/api/villas": {
            "get": {
                "tags": [
                    "Katalog Publik"
                ],
                "summary": "Daftar villa (A2)",
                "description": "Katalog villa untuk halaman listing. Tidak butuh autentikasi.\n\nKategori nonaktif ATAU yang belum punya item aktif tetap dikembalikan,\nditandai `coming_soon: true` dengan `price_from: null` — frontend\nmenampilkannya diredupkan dengan label \"Segera hadir\" (lihat kartu\nWisata Camping). Menyaringnya di server akan menghilangkan kartu itu.\n\n`price_from` dan `capacity` dihitung dari item aktif, bukan kolom\ntersimpan di kategori.",
                "operationId": "1be66933b76af35ba506c7510f579df9",
                "parameters": [
                    {
                        "name": "X-Tenant",
                        "in": "header",
                        "description": "Slug tenant. Bila kosong dipakai tenant default dari konfigurasi.",
                        "schema": {
                            "type": "string",
                            "example": "decorinna"
                        }
                    },
                    {
                        "name": "cap_min",
                        "in": "query",
                        "description": "Filter kapasitas minimum (dari dropdown \"Kapasitas tamu\").",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "cap_max",
                        "in": "query",
                        "description": "Filter kapasitas maksimum.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar villa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Villa"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tenant tidak ditemukan/tidak aktif",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/villas/{slug}": {
            "get": {
                "tags": [
                    "Katalog Publik"
                ],
                "summary": "Detail villa (A3)",
                "description": "Detail villa beserta fasilitas, item aktif (tarif weekday/weekend), dan add-on yang tertaut. Add-on disertakan di sini supaya layar pilih add-on (A7) tidak perlu panggilan terpisah.",
                "operationId": "0cc0cd54f6ea2e4fba7e1f90685d29b2",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "villa-de-corrinna"
                        }
                    },
                    {
                        "name": "X-Tenant",
                        "in": "header",
                        "schema": {
                            "type": "string",
                            "example": "decorinna"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail villa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VillaDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Villa atau tenant tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Addon": {
                "title": "Addon",
                "description": "Cerminan AddonResource.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Extra Bed"
                    },
                    "price": {
                        "type": "integer",
                        "example": 150000
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "unit": {
                        "description": "Satuan tampilan di A7: \"unit\" (Extra Bed), \"jam\" (Late Checkout), atau null untuk harga flat.",
                        "type": "string",
                        "example": "unit",
                        "nullable": true
                    },
                    "deadline_days": {
                        "description": "Batas pemesanan H-n sebelum check-in (mis. catering H-4).",
                        "type": "integer",
                        "example": 4,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "Aktif",
                        "enum": [
                            "Aktif",
                            "Nonaktif"
                        ]
                    },
                    "links": {
                        "description": "Bentuk datar dari pivot polimorfik — mengisi kolom \"Terhubung ke Item/Kategori\".",
                        "type": "array",
                        "items": {
                            "properties": {
                                "type": {
                                    "type": "string",
                                    "enum": [
                                        "category",
                                        "item"
                                    ]
                                },
                                "id": {
                                    "type": "integer"
                                },
                                "label": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        },
                        "example": [
                            {
                                "type": "category",
                                "id": 1,
                                "label": "Villa De Corrinna"
                            }
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "title": "ErrorResponse",
                "description": "Envelope response error standar.",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Terjadi kesalahan."
                    },
                    "errors": {
                        "description": "Peta field -> daftar pesan (hanya untuk error validasi).",
                        "type": "object",
                        "example": {
                            "email": [
                                "Email wajib diisi."
                            ]
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Facility": {
                "title": "Facility",
                "description": "Cerminan FacilityResource.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Private Pool"
                    },
                    "icon": {
                        "description": "Nama ikon lucide.",
                        "type": "string",
                        "example": "waves",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "Aktif",
                        "enum": [
                            "Aktif",
                            "Nonaktif"
                        ]
                    }
                },
                "type": "object"
            },
...
            "Item": {
                "title": "Item",
                "description": "Kamar/paket di dalam sebuah kategori (villa).",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Kamar Superior"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "capacity_type": {
                        "type": "string",
                        "example": "Rentang",
                        "enum": [
                            "Tetap",
                            "Rentang"
                        ]
                    },
                    "cap_min": {
                        "type": "integer",
                        "example": 8
                    },
                    "cap_max": {
                        "type": "integer",
                        "example": 12
                    },
                    "price_weekday": {
                        "description": "Rupiah sebagai integer, tidak pernah float.",
                        "type": "integer",
                        "example": 4000000
                    },
                    "price_weekend": {
                        "type": "integer",
                        "example": 4800000
                    },
                    "payment_mode": {
                        "description": "Nilai persis seperti label radio di ItemForm. ⚠ Kebijakan default per item belum final (dok 05) — backend hanya menyimpan pilihan admin.",
                        "type": "string",
                        "example": "Full Payment",
                        "enum": [
                            "Full Payment",
                            "DP + Pelunasan",
                            "Keduanya — customer memilih"
                        ]
                    },
                    "dp_minimum": {
                        "description": "Wajib bila payment_mode memakai DP.",
                        "type": "integer",
                        "example": 2000000,
                        "nullable": true
                    },
                    "requires_survey": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "string",
                        "example": "Aktif",
                        "enum": [
                            "Aktif",
                            "Nonaktif"
                        ]
                    },
                    "photos": {
                        "description": "Selalu ada; masih kosong sampai endpoint upload dibuat.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ItemPhoto"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ItemPhoto": {
                "title": "ItemPhoto",
                "description": "Cerminan ItemPhotoResource.",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "url": {
                        "type": "string",
                        "example": "/storage/items/kamar-superior-1.jpg"
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 0
                    },
                    "is_cover": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "SuccessResponse": {
                "title": "SuccessResponse",
                "description": "Envelope response sukses standar.",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Villa": {
                "title": "Villa",
                "description": "Kartu villa di katalog publik (A2).",
                "properties": {
                    "slug": {
                        "type": "string",
                        "example": "villa-de-corrinna"
                    },
                    "name": {
                        "type": "string",
                        "example": "Villa De Corrinna"
                    },
                    "icon": {
                        "type": "string",
                        "example": "house",
                        "nullable": true
                    },
                    "tagline": {
                        "type": "string",
                        "example": "Rumah utama dengan taman luas dan kolam privat.",
                        "nullable": true
                    },
                    "coming_soon": {
                        "description": "true bila kategori nonaktif ATAU belum punya item aktif. Kartu tetap dikembalikan, frontend menampilkannya diredupkan dengan label \"Segera hadir\".",
                        "type": "boolean",
                        "example": false
                    },
                    "price_from": {
                        "description": "Harga weekday termurah dari item aktif. null bila coming_soon.",
                        "type": "integer",
                        "example": 4000000,
                        "nullable": true
                    },
                    "capacity": {
                        "description": "Rentang gabungan dari item aktif — mengisi teks \"8–30 orang\".",
                        "properties": {
                            "min": {
                                "type": "integer",
                                "example": 8,
                                "nullable": true
                            },
                            "max": {
                                "type": "integer",
                                "example": 30,
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "photo": {
                        "description": "Menyusul bersama endpoint upload foto.",
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "VillaDetail": {
                "title": "VillaDetail",
                "description": "Detail villa publik (A3), termasuk add-on yang dipakai layar A7.",
                "properties": {
                    "slug": {
                        "type": "string",
                        "example": "villa-de-corrinna"
                    },
                    "name": {
                        "type": "string",
                        "example": "Villa De Corrinna"
                    },
                    "icon": {
                        "type": "string",
                        "example": "house",
                        "nullable": true
                    },
                    "tagline": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "description": "Isi bagian \"Tentang villa ini\".",
                        "type": "string",
                        "nullable": true
                    },
                    "location": {
                        "type": "string",
                        "example": "Kawasan Gerbang Gunungsari, Pamijahan, Bogor",
                        "nullable": true
                    },
                    "coming_soon": {
                        "type": "boolean",
                        "example": false
                    },
                    "price_from": {
                        "type": "integer",
                        "example": 4000000,
                        "nullable": true
                    },
                    "capacity": {
                        "properties": {
                            "min": {
                                "type": "integer",
                                "example": 8,
                                "nullable": true
                            },
                            "max": {
                                "type": "integer",
                                "example": 30,
                                "nullable": true
                            }
                        },
                        "type": "object"
                    },
                    "facilities": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Facility"
                        }
                    },
                    "items": {
                        "description": "Hanya item berstatus Aktif, diurutkan dari harga weekday termurah.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Item"
                        }
                    },
                    "addons": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Addon"
                        }
                    },
                    "photos": {
                        "description": "Menyusul bersama endpoint upload foto.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ItemPhoto"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token pribadi dari Laravel Sanctum. Kirim sebagai header: Authorization: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}